+++ CHANGELOG.md
@@ -5,11 +5,12 @@ StellarX 项目所有显著的变化都将被记录在这个文件中。
 格式基于 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 并且本项目遵循 [语义化版本](https://semver.org/lang/zh-CN/)。
 
-## [v2.0.1] - 2025 - 09 - 30
+## [v2.0.1] - 2025 - 10 - 01
 
 ### 新增
 
-- 文本框新增`setText`接口，可在外部设置文本框内容
+- `TextBox`新增`setText`API，可在外部设置文本框内容
+- `Button`新增`setButtonClick`API,允许通过外部函数修改按钮的点击状态，并执行相应的回调函数
 
 ## [v2.0.0] - 2025-09-21